Add files via upload
This document shared contains the gist of the Quadfoil project. It involves the methodologies that were followed to develop the novel drone configuration. Further, it delves into methods used such as design, flight planning/testing, CFD analysis and wind tunnel testing and discusses the results achieved with each of them.